Moved business error definitions out of business into model
