+++ source/app/blueprints/rest/case/case_timeline_routes.py
@@ -73,7 +73,7 @@ from app.blueprints.access_controls import ac_api_requires
 from app.util import add_obj_history_entry
 from app.blueprints.responses import response_error
 from app.blueprints.responses import response_success
-from app.business.errors import BusinessProcessingError
+from app.models.errors import BusinessProcessingError
 from app.business.events import events_create
 from app.business.events import events_update
 from app.business.events import events_delete
