+++ source/app/blueprints/rest/manage/manage_cases_routes.py
@@ -60,7 +60,7 @@ from app.business.cases import cases_delete
 from app.business.cases import cases_update
 from app.business.cases import cases_create
 from app.business.cases import cases_get_by_identifier
-from app.business.errors import BusinessProcessingError
+from app.models.errors import BusinessProcessingError
 from app.iris_engine.module_handler.module_handler import call_deprecated_on_preload_modules_hook
 from app.datamgmt.manage.manage_access_control_db import user_has_client_access
