+++ source/app/blueprints/rest/v2/alerts_routes/comments.py
@@ -39,7 +39,7 @@ from app.blueprints.rest.case_comments import case_comment_update
 from app.business.alerts import alerts_get
 from app.business.alerts import alerts_exists
 from app.blueprints.iris_user import iris_current_user
-from app.business.errors import ObjectNotFoundError
+from app.models.errors import ObjectNotFoundError
 
 
 class CommentsOperations:
