+++ source/app/blueprints/rest/v2/assets_routes/comments.py
@@ -35,7 +35,7 @@ from app.business.comments import comments_get_for_asset
 from app.business.comments import comments_delete_for_asset
 from app.blueprints.rest.case_comments import case_comment_update
 from app.business.assets import assets_get
-from app.business.errors import ObjectNotFoundError
+from app.models.errors import ObjectNotFoundError
 from app.schema.marshables import CommentSchema
 from app.blueprints.iris_user import iris_current_user
 from app.models.authorization import CaseAccessLevel
