+++ source/app/blueprints/rest/v2/case_routes/assets.py
@@ -36,8 +36,8 @@ from app.business.assets import assets_filter
 from app.business.assets import assets_get
 from app.business.assets import assets_update
 from app.business.assets import assets_delete
-from app.business.errors import BusinessProcessingError
-from app.business.errors import ObjectNotFoundError
+from app.models.errors import BusinessProcessingError
+from app.models.errors import ObjectNotFoundError
 from app.iris_engine.module_handler.module_handler import call_deprecated_on_preload_modules_hook
 from app.models.authorization import CaseAccessLevel
 from app.schema.marshables import CaseAssetsSchema
