+++ source/app/blueprints/rest/v2/case_routes/notes.py
@@ -35,8 +35,8 @@ from app.business.notes import notes_get
 from app.business.notes import notes_update
 from app.business.notes import notes_delete
 from app.business.cases import cases_exists
-from app.business.errors import BusinessProcessingError
-from app.business.errors import ObjectNotFoundError
+from app.models.errors import BusinessProcessingError
+from app.models.errors import ObjectNotFoundError
 from app.iris_engine.module_handler.module_handler import call_deprecated_on_preload_modules_hook
