+++ source/app/blueprints/rest/v2/case_routes/tasks.py
@@ -30,8 +30,8 @@ from app.blueprints.rest.parsing import parse_pagination_parameters
 from app.blueprints.access_controls import ac_api_return_access_denied, ac_fast_check_current_user_has_case_access
 from app.blueprints.access_controls import ac_api_requires
 from app.schema.marshables import CaseTaskSchema
-from app.business.errors import BusinessProcessingError
-from app.business.errors import ObjectNotFoundError
+from app.models.errors import BusinessProcessingError
+from app.models.errors import ObjectNotFoundError
 from app.business.tasks import tasks_create
 from app.business.tasks import tasks_get
 from app.business.tasks import tasks_update
