+++ source/app/blueprints/rest/v2/manage_routes/groups.py
@@ -33,8 +33,8 @@ from app.business.groups import groups_update
 from app.business.groups import groups_delete
 from app.models.authorization import Permissions
 from app.models.authorization import ac_flag_match_mask
-from app.business.errors import BusinessProcessingError
-from app.business.errors import ObjectNotFoundError
+from app.models.errors import BusinessProcessingError
+from app.models.errors import ObjectNotFoundError
 from app.blueprints.iris_user import iris_current_user
 from app.iris_engine.access_control.utils import ac_ldp_group_update
