+++ source/app/business/alerts.py
@@ -35,7 +35,7 @@ from app.datamgmt.alerts.alerts_db import get_related_alerts_details
 from app.iris_engine.module_handler.module_handler import call_modules_hook
 from app.iris_engine.utils.tracker import track_activity
 from app.util import add_obj_history_entry
-from app.business.errors import ObjectNotFoundError
+from app.models.errors import ObjectNotFoundError
 from app.datamgmt.manage.manage_access_control_db import user_has_client_access
