+++ source/app/business/cases.py
@@ -23,8 +23,8 @@ from app import db
 from app.logger import logger
 from app.util import add_obj_history_entry
 from app.models.models import ReviewStatusList
-from app.business.errors import BusinessProcessingError
-from app.business.errors import ObjectNotFoundError
+from app.models.errors import BusinessProcessingError
+from app.models.errors import ObjectNotFoundError
 from app.business.iocs import iocs_exports_to_json
 from app.iris_engine.module_handler.module_handler import call_modules_hook
 from app.iris_engine.utils.tracker import track_activity
