+++ source/app/business/evidences.py
@@ -19,8 +19,8 @@
 from flask_sqlalchemy.pagination import Pagination
 
 from app.blueprints.iris_user import iris_current_user
-from app.business.errors import BusinessProcessingError
-from app.business.errors import ObjectNotFoundError
+from app.models.errors import BusinessProcessingError
+from app.models.errors import ObjectNotFoundError
 from app.iris_engine.module_handler.module_handler import call_modules_hook
 from app.iris_engine.utils.tracker import track_activity
 from app.models.models import CaseReceivedFile
