+++ source/app/business/notes.py
@@ -22,9 +22,9 @@ from app import db
 from app.datamgmt.persistence_error import PersistenceError
 from app.blueprints.iris_user import iris_current_user
 from app.logger import logger
-from app.business.errors import BusinessProcessingError
-from app.business.errors import UnhandledBusinessError
-from app.business.errors import ObjectNotFoundError
+from app.models.errors import BusinessProcessingError
+from app.models.errors import UnhandledBusinessError
+from app.models.errors import ObjectNotFoundError
 from app.datamgmt.case.case_notes_db import search_notes_in_case
 from app.datamgmt.case.case_notes_db import get_note
 from app.datamgmt.case.case_notes_db import update_note_revision
