+++ source/app/business/notes_directories.py
@@ -19,7 +19,7 @@
 from app import db
 from app.iris_engine.utils.tracker import track_activity
 from app.models.models import NoteDirectory
-from app.business.errors import ObjectNotFoundError
+from app.models.errors import ObjectNotFoundError
 from app.datamgmt.case.case_notes_db import get_directory
 from app.datamgmt.case.case_notes_db import delete_directory
 from app.datamgmt.case.case_notes_db import paginate_notes_directories
