+++ source/app/business/reports/reporter.py
@@ -22,7 +22,7 @@ import logging as log
 import os
 from datetime import datetime
 
-from app.business.errors import BusinessProcessingError
+from app.models.errors import BusinessProcessingError
 from app.blueprints.iris_user import iris_current_user
 from docx_generator.docx_generator import DocxGenerator
 from docx_generator.exceptions import rendering_error
