+++ source/app/business/tasks.py
@@ -33,8 +33,8 @@ from app.iris_engine.module_handler.module_handler import call_modules_hook
 from app.iris_engine.utils.tracker import track_activity
 from app.models.models import CaseTasks
 from app.models.pagination_parameters import PaginationParameters
-from app.business.errors import BusinessProcessingError
-from app.business.errors import ObjectNotFoundError
+from app.models.errors import BusinessProcessingError
+from app.models.errors import ObjectNotFoundError
 
 
 def tasks_delete(task: CaseTasks):
